+++ src/Form/TutorialType.php
@@ -1,0 +1,56 @@
+<?php
+
+namespace App\Form;
+
+use App\Entity\Tutorial;
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\FileType;
+use Symfony\Component\Form\Extension\Core\Type\SubmitType;
+use Symfony\Component\Form\Extension\Core\Type\TextareaType;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+
+class TutorialType extends AbstractType
+{
+    public function buildForm(FormBuilderInterface $builder, array $options): void
+    {
+        $builder
+            ->add('image', FileType::class, [
+                'mapped' => false,
+                'required' => $options['is_require'],
+                'attr' => [
+                    'class' => 'form-control',
+                    'placeholder' => 'Choisir une image'
+                ],
+                'label' => 'Image associée',
+                'label_attr' => [
+                    'class' => 'form-label fw-bold'
+                ]
+
+            ])
+            ->add('description', TextareaType::class, [
+                'attr' => [
+                    'class' => 'form-control',
+                    'placeholder' => "Description de l'image",
+                ],
+                'label' => 'Description',
+                'label_attr' => [
+                    'class' => 'form-label fw-bold'
+                ]
+            ])
+            ->add('submit', SubmitType::class, [
+                'attr' => [
+                    'class' => 'btn btn-primary'
+                ],
+                'label' => 'Ajouter'
+            ]);
+    }
+
+    public function configureOptions(OptionsResolver $resolver): void
+    {
+        $resolver->setDefaults([
+            'data_class' => Tutorial::class,
+            'is_require' => false,
+        ]);
+    }
+}